// frontend/types/supabase.ts
export type Database = {
  // Allows automatically instantiating createClient with the right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
      incidents: {
        Row: {
          age: number | null
          agency_ids: string | null
          armed_with: string | null
          body_camera: boolean | null
          city: string | null
          county: string | null
          date: string | null
          flee_status: string | null
          gender: string | null
          id: number | null
          latitude: number | null
          location_precision: string | null
          longitude: number | null
          name: string | null
          race: string | null
          race_source: string | null
          state: string | null
          threat_type: string | null
          was_mental_illness_related: boolean | null
        }
        Insert: {
          age?: number | null
          agency_ids?: string | null
          armed_with?: string | null
          body_camera?: boolean | null
          city?: string | null
          county?: string | null
          date?: string | null
          flee_status?: string | null
          gender?: string | null
          id?: number | null
          latitude?: number | null
          location_precision?: string | null
          longitude?: number | null
          name?: string | null
          race?: string | null
          race_source?: string | null
          state?: string | null
          threat_type?: string | null
          was_mental_illness_related?: boolean | null
        }
        Update: {
          age?: number | null
          agency_ids?: string | null
          armed_with?: string | null
          body_camera?: boolean | null
          city?: string | null
          county?: string | null
          date?: string | null
          flee_status?: string | null
          gender?: string | null
          id?: number | null
          latitude?: number | null
          location_precision?: string | null
          longitude?: number | null
          name?: string | null
          race?: string | null
          race_source?: string | null
          state?: string | null
          threat_type?: string | null
          was_mental_illness_related?: boolean | null
        }
        Relationships: []
      }
      shootings: {
        Row: {
          age: number | null
          armed_with: string | null
          body_camera: boolean | null
          city: string | null
          date: string | null
          flee_status: string | null
          gender: string | null
          id: number
          name: string | null
          race: string | null
          state: string | null
          threat_type: string | null
        }
        Insert: {
          age?: number | null
          armed_with?: string | null
          body_camera?: boolean | null
          city?: string | null
          date?: string | null
          flee_status?: string | null
          gender?: string | null
          id?: number
          name?: string | null
          race?: string | null
          state?: string | null
          threat_type?: string | null
        }
        Update: {
          age?: number | null
          armed_with?: string | null
          body_camera?: boolean | null
          city?: string | null
          date?: string | null
          flee_status?: string | null
          gender?: string | null
          id?: number
          name?: string | null
          race?: string | null
          state?: string | null
          threat_type?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never